update
- Added Master-Node-Interface library.
- Next line of action is to test communication between Master and Node

+++ Node/TDS.h
@@ -14,5 +14,5 @@ class TDS
   public:
     TDS(uint8_t signalPin,uint8_t vccControlPin,uint8_t gndControlPin,
         float refVoltage = 5.0,int adcResolution = 1024);
-    float GetValue(float temperature);
+    int GetValue(float temperature);
 };
